Documentation and more
Adjusted some scripts to have better docs. Fiddled with movement and animation controller scripts in pursuit of a solution to the moving platform issue. Added a high friction sphere to the base of the character but left the wider low friction buffer above it. Also redid the lighting for some reason.