Got Jumping working properly
Jumping is now in its own component and works pretty well. You can either use default movement in the air with changed speed and acceleration, or turn off default movement while in the air and use a whole new function to do it.

Also removed the old prefab since it was severely deprecated. Will need to add some Demo Prefabs for the peeps.